BT: Handle service_IDs properly[1/2]
 - when UBSan is enabled, observed tombstone in bluetooth
   as 32bit integer is not enough for BTA_MAX_SERVICE_ID(34)

 - This patch will handle service_IDs properly.

CRs-Fixed: 2353784

+++ bta/include/bta_api.h
@@ -139,7 +139,7 @@ typedef uint8_t tBTA_SERVICE_ID;
 #define BTA_ALL_SERVICE_MASK 0x7FFFFFFF  /* All services supported by BTA. */
 #define BTA_USER_SERVICE_MASK 0x80000000 /* Message Notification Profile */
 
-typedef uint32_t tBTA_SERVICE_MASK;
+typedef uint64_t tBTA_SERVICE_MASK;
 
 /* extended service mask, including mask with one or more GATT UUID */
 typedef struct {
